fixup! fixup! KTOR-3721 Verify StatusPages handle exception
